Rework consul_template::watch to take config_hash
This maintains backwards compatibility with the named params, but defaults their
values to undef. All of the validations and behaviors have been maintained. The
advantage of this approach is that arbitrary configuration can be passed through
to the config file similar to how the consul module works. This way each new
param doesn't have to be individually added before the module will support it.